requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django==4.2.6
django-bootstrap4==1.1.1
django-debug-toolbar==4.2.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
